Removed Commented Ruby Code
Ruby code does not like to be heavily commented. I would move one
comment and all the rest comments would need to be moved as well.
I just deleted the code we were removing to make the 'rake' tests
pass. We can pull the file(s) from the community code if we need
to redeploy this code.